hw_3_3.orders. Рефакторинг. Уточнение возвращаемых значений

+++ otus/www/classes/Dto/OrderDtoBuilder.php
@@ -29,7 +29,7 @@ class OrderDtoBuilder
         return $this;
     }
 
-    public function setCost (int $cost)
+    public function setCost (float $cost)
     {
         $this->cost = $cost;
         return $this;
